requirements.txt
django-crispy-forms
crispy-bootstrap4
django-allauth
django-allauth[socialaccount]
django-environ